feat: GTE classification head
