+++ backends/candle/tests/test_bert.rs
@@ -13,7 +13,7 @@ fn test_mini() -> Result<()> {
     let tokenizer = load_tokenizer(&model_root)?;
 
     let backend = CandleBackend::new(
-        model_root,
+        &model_root,
         "float32".to_string(),
         ModelType::Embedding(Pool::Mean),
     )?;
@@ -73,7 +73,7 @@ fn test_mini_pooled_raw() -> Result<()> {
     let tokenizer = load_tokenizer(&model_root)?;
 
     let backend = CandleBackend::new(
-        model_root,
+        &model_root,
         "float32".to_string(),
         ModelType::Embedding(Pool::Cls),
     )?;
@@ -142,7 +142,7 @@ fn test_emotions() -> Result<()> {
     let model_root = download_artifacts("SamLowe/roberta-base-go_emotions", None)?;
     let tokenizer = load_tokenizer(&model_root)?;
 
-    let backend = CandleBackend::new(model_root, "float32".to_string(), ModelType::Classifier)?;
+    let backend = CandleBackend::new(&model_root, "float32".to_string(), ModelType::Classifier)?;
 
     let input_batch = batch(
         vec![
@@ -192,7 +192,7 @@ fn test_bert_classification() -> Result<()> {
     let model_root = download_artifacts("ibm/re2g-reranker-nq", Some("refs/pr/3"))?;
     let tokenizer = load_tokenizer(&model_root)?;
 
-    let backend = CandleBackend::new(model_root, "float32".to_string(), ModelType::Classifier)?;
+    let backend = CandleBackend::new(&model_root, "float32".to_string(), ModelType::Classifier)?;
 
     let input_single = batch(
         vec![tokenizer
