+++ backends/candle/tests/test_flash_bert.rs
@@ -19,7 +19,7 @@ fn test_flash_mini() -> Result<()> {
     let tokenizer = load_tokenizer(&model_root)?;
 
     let backend = CandleBackend::new(
-        model_root,
+        &model_root,
         "float16".to_string(),
         ModelType::Embedding(Pool::Mean),
     )?;
@@ -83,7 +83,7 @@ fn test_flash_mini_pooled_raw() -> Result<()> {
     let tokenizer = load_tokenizer(&model_root)?;
 
     let backend = CandleBackend::new(
-        model_root,
+        &model_root,
         "float16".to_string(),
         ModelType::Embedding(Pool::Cls),
     )?;
@@ -156,7 +156,7 @@ fn test_flash_emotions() -> Result<()> {
     let model_root = download_artifacts("SamLowe/roberta-base-go_emotions", None)?;
     let tokenizer = load_tokenizer(&model_root)?;
 
-    let backend = CandleBackend::new(model_root, "float16".to_string(), ModelType::Classifier)?;
+    let backend = CandleBackend::new(&model_root, "float16".to_string(), ModelType::Classifier)?;
 
     let input_batch = batch(
         vec![
@@ -210,7 +210,7 @@ fn test_flash_bert_classification() -> Result<()> {
     let model_root = download_artifacts("ibm/re2g-reranker-nq", Some("refs/pr/3"))?;
     let tokenizer = load_tokenizer(&model_root)?;
 
-    let backend = CandleBackend::new(model_root, "float16".to_string(), ModelType::Classifier)?;
+    let backend = CandleBackend::new(&model_root, "float16".to_string(), ModelType::Classifier)?;
 
     let input_single = batch(
         vec![tokenizer
