+++ backends/candle/tests/test_flash_gte.rs
@@ -15,7 +15,7 @@ fn test_flash_gte() -> Result<()> {
     let tokenizer = load_tokenizer(&model_root)?;
 
     let backend = CandleBackend::new(
-        model_root,
+        &model_root,
         "float16".to_string(),
         ModelType::Embedding(Pool::Cls),
     )?;
@@ -59,13 +59,10 @@ fn test_flash_gte() -> Result<()> {
     any(feature = "flash-attn", feature = "flash-attn-v1")
 ))]
 fn test_flash_gte_classification() -> Result<()> {
-    let model_root = download_artifacts(
-        "Alibaba-NLP/gte-multilingual-reranker-base",
-        Some("refs/pr/11"),
-    )?;
+    let model_root = download_artifacts("Alibaba-NLP/gte-multilingual-reranker-base", None)?;
     let tokenizer = load_tokenizer(&model_root)?;
 
-    let backend = CandleBackend::new(model_root, "float16".to_string(), ModelType::Classifier)?;
+    let backend = CandleBackend::new(&model_root, "float16".to_string(), ModelType::Classifier)?;
 
     let input_single = batch(
         vec![tokenizer
