+++ backends/candle/tests/test_flash_mistral.rs
@@ -15,7 +15,7 @@ fn test_flash_mistral() -> Result<()> {
     let tokenizer = load_tokenizer(&model_root)?;
 
     let backend = CandleBackend::new(
-        model_root,
+        &model_root,
         "float16".to_string(),
         ModelType::Embedding(Pool::Mean),
     )?;
