+++ backends/candle/tests/test_nomic.rs
@@ -12,7 +12,7 @@ fn test_nomic_small() -> Result<()> {
     let tokenizer = load_tokenizer(&model_root)?;
 
     let backend = CandleBackend::new(
-        model_root,
+        &model_root,
         "float32".to_string(),
         ModelType::Embedding(Pool::Mean),
     )?;
